+++ test/test_overlap_turbomole.py
@@ -20,6 +20,15 @@ path_hdf5 = 'test/test_files/test.hdf5'
 path_MO = 'test/test_files/aomix_ethylene.in'
 path_xyz = 'test/test_files/ethylene_au.xyz'
 
+# Path to Nodes in the HDF5 file
+path_ethylene = join('/turbomole', 'test', 'ethylene')
+path_es = join(path_ethylene, 'eigenvalues')
+path_css = join(path_ethylene, 'coefficients')
+
+# Orbital info
+number_of_orbs = 36
+number_of_orb_funs = 38
+
 
 def dump_MOs_coeff(handle_hdf5, path_es, path_css, number_of_orbs,
                    number_of_orb_funs):
@@ -57,16 +66,12 @@ def test_store_MO_h5():
     """
     test if the MO are stored in the HDF5 format
     """
-    path = join('/turbomole', 'test', 'ethylene')
-    path_es = join(path, 'eigenvalues')
-    path_css = join(path, 'coefficients')
-    number_of_orbs = 36
-    number_of_orb_funs = 38
 
     with h5py.File(path_hdf5, chunks=True) as f5:
-        path_es, path_css = dump_MOs_coeff(f5, path_es, path_css,
-                                           number_of_orbs, number_of_orb_funs)
-        if not(f5[path_es] and f5[path_css]):
+        path_eigenvals, path_coeffs = dump_MOs_coeff(f5, path_es, path_css,
+                                                     number_of_orbs,
+                                                     number_of_orb_funs)
+        if not(f5[path_eigenvals] and f5[path_coeffs]):
             assert False
 
 
@@ -81,13 +86,6 @@ def test_overlap():
     mol = readXYZ(path_xyz)
     labels = [at.symbol for at in mol]
 
-    path = join('/turbomole', 'test', 'ethylene')
-    path_es = join(path, 'eigenvalues')
-    path_css = join(path, 'coefficients')
-
-    number_of_orbs = 36
-    number_of_orb_funs = 38
-
     # Build the Conctracted Gauss Functions
     dictCGFs = create_dict_CGFs(path_hdf5, basis, mol,
                                 package_name='turbomole',
@@ -95,8 +93,9 @@ def test_overlap():
     cgfsN = [dictCGFs[l] for l in labels]
 
     with h5py.File(path_hdf5, chunks=True) as f5:
-        path_es, path_css = dump_MOs_coeff(f5, path_es, path_css,
-                                           number_of_orbs, number_of_orb_funs)
+        path_eigenvals, path_coeffs = dump_MOs_coeff(f5, path_es, path_css,
+                                                     number_of_orbs,
+                                                     number_of_orb_funs)
         trr = f5[path_css].value
 
     dim = sum(len(xs) for xs in cgfsN)
